requirements.txt
playwright>=1.42.0
pytest>=7.0.0
allure-pytest>=2.13.2
openpyxl>=3.1.2
pytest-playwright>=0.3.0
pytest-xdist>=3.3.1
